Refactoring and enabling ruby dependencies
