+++ tests/integration/testdata/sync/infra/after/Python/apigateway-after/definition.json
