+++ tests/integration/testdata/sync/infra/after/Python/apigateway2-after/definition.yaml
