+++ tests/integration/testdata/sync/infra/after/Python/statemachine-after/function.asl.json
