+++ tests/integration/testdata/sync/infra/before/Python/apigateway-before/definition.json
