+++ tests/integration/testdata/sync/infra/before/Python/apigateway2-before/definition.yaml
