+++ tests/integration/testdata/sync/infra/before/Python/function-before/requirements.txt
