+++ tests/integration/testdata/sync/infra/before/Python/statemachine-before/function.asl.json
