+++ tests/integration/testdata/sync/infra/before/Ruby/function-before/Gemfile
