+++ tests/integration/testdata/sync/infra/before/Ruby/function-before/app.rb
@@ -1,0 +1,22 @@
+require 'httparty'
+require 'json'
+require 'layer_test'
+
+def lambda_handler(event:, context:)
+  # Sample pure Lambda function that returns a message and a location
+
+  begin
+    response = HTTParty.get('http://checkip.amazonaws.com/')
+  rescue HTTParty::Error => error
+    puts error.inspect
+    raise error
+  end
+
+  {
+    statusCode: 200,
+    body: {
+      message: "#{layer_test()+1}",
+      location: response.body
+    }.to_json
+  }
+end
