+++ tests/integration/testdata/sync/infra/before/Ruby/layer-before/Gemfile
