+++ tests/integration/testdata/sync/infra/template-python-after.yaml
@@ -9,7 +9,7 @@ Resources:
   HelloWorldFunction:
     Type: AWS::Serverless::Function
     Properties:
-      CodeUri: after/Python/function/
+      CodeUri: after/Python/function-after/
       Handler: app.lambda_handler
       Runtime: python3.7
       Layers:
@@ -20,20 +20,20 @@ Resources:
     Type: AWS::Serverless::Api
     Properties:
       StageName: prod
-      DefinitionUri: after/Python/apigateway/definition.json
+      DefinitionUri: after/Python/apigateway-after/definition.json
 
   HelloWorldApiV2:
     Type: AWS::Serverless::HttpApi
     Properties:
       StageName: beta
-      DefinitionUri: after/Python/apigateway2/definition.yaml
+      DefinitionUri: after/Python/apigateway2-after/definition.yaml
 
   HelloWorldLayer:
     Type: AWS::Serverless::LayerVersion
     Properties:
       LayerName: HelloWorldLayer
       Description: Hello World Layer
-      ContentUri: after/Python/layer/
+      ContentUri: after/Python/layer-after/
       CompatibleRuntimes:
         - python3.7
     Metadata:
@@ -42,7 +42,7 @@ Resources:
   HelloStepFunction:
     Type: AWS::Serverless::StateMachine
     Properties:
-      DefinitionUri: after/Python/statemachine/function.asl.json
+      DefinitionUri: after/Python/statemachine-after/function.asl.json
       Policies:
         - LambdaInvokePolicy:
             FunctionName: !Ref HelloWorldFunction
