+++ tests/integration/testdata/sync/infra/template-python-before.yaml
@@ -9,7 +9,7 @@ Resources:
   HelloWorldFunction:
     Type: AWS::Serverless::Function
     Properties:
-      CodeUri: before/Python/function/
+      CodeUri: before/Python/function-before/
       Handler: app.lambda_handler
       Runtime: python3.7
       Layers:
@@ -20,20 +20,20 @@ Resources:
     Type: AWS::Serverless::Api
     Properties:
       StageName: prod
-      DefinitionUri: before/Python/apigateway/definition.json
+      DefinitionUri: before/Python/apigateway-before/definition.json
 
   HelloWorldApiV2:
     Type: AWS::Serverless::HttpApi
     Properties:
       StageName: beta
-      DefinitionUri: before/Python/apigateway2/definition.yaml
+      DefinitionUri: before/Python/apigateway2-before/definition.yaml
 
   HelloWorldLayer:
     Type: AWS::Serverless::LayerVersion
     Properties:
       LayerName: HelloWorldLayer
       Description: Hello World Layer
-      ContentUri: before/Python/layer/
+      ContentUri: before/Python/layer-before/
       CompatibleRuntimes:
         - python3.7
     Metadata:
@@ -42,7 +42,7 @@ Resources:
   HelloStepFunction:
     Type: AWS::Serverless::StateMachine
     Properties:
-      DefinitionUri: before/Python/statemachine/function.asl.json
+      DefinitionUri: before/Python/statemachine-before/function.asl.json
       Policies:
         - LambdaInvokePolicy:
             FunctionName: !Ref HelloWorldFunction
